refactor logger dir
