+++ src/logger/CMakeLists.txt
@@ -1,9 +1,6 @@
 
-add_library(logger SHARED "Logger.cpp")
+add_library(logger SHARED "Include/Src/Logger.cpp")
 
 add_executable(example "Example.cpp")
 
 target_link_libraries(example logger)
-
-add_executable(con_test "Thr_test.cpp")
-target_link_libraries(con_test logger)
