+++ src/logger/Include/Src/SharedInclude.h
